fix claimable in calculation
